+++ R/flextable.R
@@ -1,15 +1,4 @@
 
-
-
-#' @rdname as_flextable
-#'
-#' @export
-as_FlexTable <- function(x, ...) {
-  .Deprecated(new = "as_flextable", package = "huxtable")
-  as_flextable(x, ...)
-}
-
-
 #' Convert a huxtable for Word/Powerpoint
 #'
 #' Huxtables can be converted to [flextable::flextable()] objects, for use in Word and Powerpoint documents.
@@ -26,17 +15,15 @@ as_FlexTable <- function(x, ...) {
 #' from rmarkdown `word_document` and/or `powerpoint_presentation` documents. (Powerpoint
 #' presentations require pandoc version >= 2.4.0.)
 #'
-#' `as_FlexTable` is deprecated and calls `as_flextable` with a warning.
-#'
 #' Properties are supported, with the following exceptions:
-
 #' * Rotation of 0, 90 or 270 is supported.
 #' * Non-numeric widths and heights are not supported. Table heights are treated
 #'   as a proportion of 9 inches; table widths are treated as a proportion of 6
 #'   inches. So e.g. `height(ht) <- 0.5` will give a height of 4.5 inches.
 #' * Table wrap and table position are not supported.
 #' * Border style "double" is not supported and becomes "solid".
-#
+#' * Captions are supported with recent versions of flextable, but not
+#'   [caption_pos()] or [caption_width()].
 #'
 #' @section Challenge:
 #'