+++ man-roxygen/cell-property-usage.R
@@ -3,5 +3,5 @@
 #' @usage
 #' <%= attr_name %>(ht)
 #' <%= attr_name %>(ht) <- value
-#' set_<%= attr_name %>(ht, row, col, value, byrow = FALSE)
+#' set_<%= attr_name %>(ht, row, col, value)
 #' map_<%= attr_name %>(ht, row, col, fn)